+++ src/createAndSendMessage.ts
